backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg[binary]==3.2.10
pydantic==2.9.2
python-dotenv==1.0.1
email-validator==2.2.0